extract index into project structure
