+++ src/cli/index.js
@@ -1,4 +1,3 @@
-const game = require('../game')
 const makeMove = require('./makeMove')
 
 const start = (options, io) => {
@@ -10,17 +9,6 @@ const start = (options, io) => {
   write(messages.instructions)
 
   makeMove(options, io)
-}
-
-if (require.main === module) {
-  const io = {
-    input: process.stdin,
-    output: process.stdout,
-    write: console.log,
-  }
-
-  const options = game.init()
-  start(options, io)
 }
 
 module.exports = start
